+++ flowtorch/distributions/flow.py
@@ -3,7 +3,6 @@
 from typing import Any, Dict, Optional, Union, Iterator
 
 import flowtorch
-
 import torch
 import torch.distributions as dist
 from torch import Tensor
