+++ tests/test_bijector.py
@@ -3,7 +3,6 @@ import warnings
 
 import flowtorch.bijectors as bijectors
 import flowtorch.parameters as params
-
 import numpy as np
 import pytest
 import torch
